+++ shared/box.cpp
@@ -1,6 +1,6 @@
 #include "box.h"
 
-box::box(level *lvl, char abox_type, vec *pos, uint ahealth)
+box::box(level *lvl, char abox_type, vec *pos, float ahealth)
 	: actor(lvl, ACTOR_TYPE_BOX, pos, NULL)
 {
 	health = ahealth;
@@ -8,7 +8,7 @@ box::box(level *lvl, char abox_type, vec *pos, uint ahealth)
 }
 
 
-box::box(level *lvl, uint actor_id, char abox_type, vec *pos, uint ahealth)
+box::box(level *lvl, uint actor_id, char abox_type, vec *pos, float ahealth)
 	: actor(lvl, actor_id, ACTOR_TYPE_BOX, pos, NULL)
 {
 	health = ahealth;
